+++ apps/native-component-list/src/screens/UI/UIScreen.android.tsx
@@ -147,6 +147,14 @@ export const UIScreens = [
     },
   },
   {
+    name: 'Community PagerView replacement',
+    route: 'ui/community-pager-view',
+    options: {},
+    getComponent() {
+      return optionalRequire(() => require('./CommunityPagerViewScreen'));
+    },
+  },
+  {
     name: 'Switch component',
     route: 'ui/switch',
     options: {},