+++ apps/native-component-list/src/screens/UI/UIScreen.tsx
@@ -34,6 +34,14 @@ export const UIScreens = [
       return optionalRequire(() => require('./CommunitySliderScreen'));
     },
   },
+  {
+    name: 'Community PagerView replacement',
+    route: 'ui/community-pager-view',
+    options: {},
+    getComponent() {
+      return optionalRequire(() => require('./CommunityPagerViewScreen'));
+    },
+  },
 ];
 
 export default function UIScreen() {